Conflicts:
	Trombone/.classpath
	Trombone/AndroidManifest.xml
	Trombone/bin/AndroidManifest.xml
	Trombone/gen/com/example/trombone/R.java
	Trombone/res/layout/activity_main_ui.xml
	Trombone/res/values/colors.xml
	Trombone/res/values/strings.xml
	Trombone/src/com/example/trombone/MainUIActivity.java

+++ Trombone/src/classes/Memo.java
@@ -4,7 +4,7 @@ public class Memo {
 	public Memo() {
 		super();
 	}
-	public Memo(int id, int x, int y, int opacity, int page, String content,
+	public Memo(int id, float x, float y, int opacity, int page, String content,
 			int musicsheet_id) {
 		super();
 		this.id = id;
@@ -21,13 +21,13 @@ public class Memo {
 	public void setId(int id) {
 		this.id = id;
 	}
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 	public void setX(int x) {
 		this.x = x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public void setY(int y) {
@@ -59,9 +59,9 @@ public class Memo {
 	}
 	
 	private int id;
-	private int x;
-	private int y;
-	private int opacity;  // 1~100
+	private float x;
+	private float y;
+	private int opacity;  // 1~255
 	private int page;
 	private String content;
 	private int musicsheet_id;
